Overhauled the QML code.
Broke the code for each screen into a different qml file. Created ShowerHider and Focuser to streamline the process of showing, hiding, and focusing elements. Implemented it for all the screens. Also fixed an issue with K being useable to move through photos (although it would put 'continue' into the data instead of actual data).